+++ src/hooks/api/useContentUpdate.js
@@ -1,8 +1,10 @@
 import { useMutation } from 'react-query';
 import { useOkapiKy } from '@folio/stripes/core';
+import { useErrorMessages } from '../useErrorMessages';
 
 export const useContentUpdate = ({ id }) => {
   const ky = useOkapiKy();
+  const { showErrorMessage } = useErrorMessages();
 
   const { data, mutateAsync: contentUpdate, isLoading } = useMutation({
     mutationFn: ({ contentUpdates }) => {
@@ -10,6 +12,7 @@ export const useContentUpdate = ({ id }) => {
         json: contentUpdates,
       });
     },
+    onSuccess: showErrorMessage,
   });
 
   return { contentUpdate, isLoading, data };